Chore: Refactor renderer package to improve type safety and limit dependence on lib/Resource.ts
