+++ packages/app-desktop/gui/NoteEditor/utils/types.ts
@@ -2,7 +2,7 @@ import AsyncActionQueue from '@joplin/lib/AsyncActionQueue';
 import { ToolbarButtonInfo } from '@joplin/lib/services/commands/ToolbarButtonUtils';
 import { PluginStates } from '@joplin/lib/services/plugins/reducer';
 import { MarkupLanguage } from '@joplin/renderer';
-import { RenderResult, RenderResultPluginAsset } from '@joplin/renderer/MarkupToHtml';
+import { RenderResult, RenderResultPluginAsset } from '@joplin/renderer/types';
 import { MarkupToHtmlOptions } from './useMarkupToHtml';
 import { Dispatch } from 'redux';
 import { ProcessResultsRow } from '@joplin/lib/services/search/SearchEngine';
